pyproject.toml
[tool.poetry]
name = "mossnobphoto-bot"
version = "0.1.0"
description = ""
authors = ["John Doe <ifadeev@more.tv>"]
readme = "README.md"
packages = [{include = "mossnobphoto_bot"}]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.21.0"
aiogram = "^2.23.1"
redis = "^4.3.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
